support for sequencing constructors/methods with or without primitive type parameters, return primitive type and more test
git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@430 76366958-4244-0410-ad5e-bbfabb93f86b

+++ extensions/dna-sequencer-java/src/test/workspace/projectX/src/org/acme/MySource.java
@@ -56,7 +56,7 @@ public class MySource {
         this.j = j;
     }
 
-    public void doSomething(int p1, int p2) {
+    public void doSomething(int p1, double p2) {
         l = new ArrayList<String>();
         l.add("N1");
     }
